Extra permission check on loop
Without this extra check, sub commands are displayed even if the player doesn't have permission to use the command.

This may be as intended. However, if it is, then it contradicts the page count having a permission check, in a previous loop. This can lead to confusing page counts. For example:

If a player doesn't have any permissions, they would see a page count of 0/0. However, without this check, they would be able to see all of the subcommands, leading to confusing page counts (as said before), and allowing them to see any and all commands (for at least the first page. I didn't check to see if it was possible to see commands of a second page, before just fixing the issue).