Currently, when adding Subcommands (and maybe some aliases), an assertion is done to make sure that helpList is still null.
This is because, once the helpList is created and sorted, any commands added to the subcommandMap wouldn't be visible. This is a rare case (potentially impossible), that would only happen if a CommandSender could somehow generate the helpList before the subCommands are completely implemented.

Now, having an assertion is good, because it allows for a better understanding of an error when it occurs. However, in this specific situation, I believe that simply taking the time to build the helpList again would be more beneficial than forcing the console to hit an exception. It doesn't use enough resources to bother mentioning, and completely solves this extremely rare occurrence.